backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
environs
django-cors-headers
psycopg2-binary
PyGithub
pytest
pytest-django
factory-boy